pyproject.toml
[tool.poetry]
name = "telegram_notification_bot"
version = "0.1.0"
description = ""
authors = ["WildElaeagnus <54314123+akorzunin@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.9"
python-telegram-bot = "^13.11"
python-dotenv = "^0.20.0"
loguru = "^0.6.0"
asyncio = "^3.4.3"
fastapi = "^0.75.2"
aiogram = "^2.20"
uvicorn = "^0.17.6"
nest-asyncio = "^1.5.5"
python-binance = "^1.0.16"
pandas = "^1.4.2"

[tool.poetry.dev-dependencies]
mypy = "^0.942"
isort = "^5.10.1"
ipykernel = "^6.13.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
